This fixes two incorrect function prototypes in headers.
display.h lists the first two parameters as s1 and s2, whereas the implementation lists them as s0 and s1.
I changed the header to be s0 and s1, seeing as the other parameters start at 0 too.

table.h lists the first two parameters as col then row, but the implementation lists them as row then col.
Judging by set_item, the order should be row then col, so I changed it to that.

Not essential fixes, but the row/col thing could be confusing for people who rely on tooltips.

Only tested on Windows, but it shouldn't affect other platforms.